Attempt to track actual wall time in async_wrapper
Previously the async wrapper had two issues with its timeout tracking.
The first was that it did one sleep without accounting for that sleep in
the remaining time calculations so it would potentially sleep one step
too long. The second is that it assumes that time.sleep() always sleeps
for exactly the amount of time passed to it, but it may sleep for a
shorter time if any signals are received by the process.

Address the first issue by only sleeping in the while loop so that every
sleep is accounted for in the remaining time calculations. Address the
second issue by checking the current time.time() on each iteration to
determine the actual remaining wall time instead of the assumed value
based on step.